refactor: resolve clippy::module_inception
